Compile.
NSInteger is int/long on 32/64 bit.

+++ src/gui/kernel/qt_cocoa_helpers_mac.mm
@@ -648,7 +648,7 @@ bool qt_dispatchKeyEventWithCocoa(void * /*NSEvent * */ keyEvent, QWidget *widge
 }
 #endif
 
-Qt::MouseButton cocoaButton2QtButton(long buttonNum)
+Qt::MouseButton cocoaButton2QtButton(NSInteger buttonNum)
 {
     if (buttonNum == 0)
         return Qt::LeftButton;
